requirements-pi.txt
# Core dependencies
PyYAML>=6.0 # For configuration loading/saving
# Note: Pillow is NOT included here unless specifically needed on Pi for other tasks
gradio>=4.0 # For the web UI

# Raspberry Pi specific dependencies
# The 'rgbmatrix' package comes from the rpi-rgb-led-matrix library build process
# Installation typically involves cloning the repo and running 'make build-python'
# There isn't usually a direct PyPI package for the core library bindings.
# Add a comment indicating how to install it.
# rgbmatrix # Install via 'sudo make install-python' in rpi-rgb-led-matrix repo

# Add other Pi-specific dependencies here if needed
